chore(process-cooling): move filter outputs pipe
